requirements.txt
sphinx-automodapi
sphinx_autodoc_typehints
numpydoc
pytest
git+https://github.com/noaa-ocs-hydrography/drivers.git#egg=hstb.drivers
git+https://github.com/noaa-ocs-hydrography/shared.git#egg=hstb.shared
git+https://github.com/noaa-ocs-hydrography/vyperdatum.git#egg=vyperdatum
git+https://github.com/noaa-ocs-hydrography/bathygrid.git#egg=bathygrid